feat: add entry point for weather
